POLIO-127: refactor according to review comments
