+++ plugins/polio/js/src/pages/IM/utils.js
@@ -1,8 +1,4 @@
-import React from 'react';
-import CheckIcon from '@material-ui/icons/Check';
-import { Box } from '@material-ui/core';
 import { IM_PASS, IM_FAIL, IM_WARNING } from './constants';
-import MESSAGES from '../../constants/messages';
 
 export const determineStatusForDistrict = district => {
     if (!district) return null;
@@ -11,13 +7,6 @@ export const determineStatusForDistrict = district => {
     if (ratio >= 95) return IM_PASS;
     if (ratio > 89 && ratio < 95) return IM_WARNING;
     return IM_FAIL;
-};
-
-const applyStatusColor = status => {
-    if (status === IM_PASS) return { color: 'green' };
-    if (status === IM_FAIL) return { color: 'red' };
-    if (status === IM_WARNING) return { color: 'orange' };
-    return null;
 };
 
 export const getImStatsForRound = (lqasData, campaign, round) => {
@@ -31,48 +20,4 @@ export const getImStatsForRound = (lqasData, campaign, round) => {
     const failed = allStatuses.filter(status => status === IM_FAIL);
 
     return [totalEvaluated, passed, failed, disqualified];
-};
-
-export const imTableColumns = formatMessage => {
-    return [
-        {
-            Header: formatMessage(MESSAGES.districtName),
-            accessor: 'name',
-            sortable: true,
-        },
-        {
-            Header: formatMessage(MESSAGES.childrenMarked),
-            accessor: 'total_child_fmd',
-            sortable: false,
-        },
-        {
-            Header: formatMessage(MESSAGES.childrenChecked),
-            accessor: 'total_child_checked',
-            sortable: false,
-        },
-        {
-            Header: formatMessage(MESSAGES.status),
-            accessor: 'status',
-            sortable: false,
-            Cell: settings => (
-                <span style={applyStatusColor(settings.row.original.status)}>
-                    {formatMessage(MESSAGES[settings.row.original.status])}
-                </span>
-            ),
-        },
-        {
-            Header: formatMessage(MESSAGES.districtFound),
-            accessor: 'district',
-            sortable: true,
-            Cell: settings => {
-                if (settings.row.original.district)
-                    return (
-                        <Box>
-                            <CheckIcon />
-                        </Box>
-                    );
-                return null;
-            },
-        },
-    ];
 };
